fix: move latest endpoint inside of trajectories and change responses in  message
