+++ src/models/trajectoriesModel.ts
@@ -19,3 +19,24 @@ export const getTrajectories = async (taxiId: number, date: string) => {
 
     return results;   
 }
+
+//--------------------Latest--------------------------------------
+export const getLatestTrajectories = async () => {
+
+    return await prisma.$queryRaw<Trajectory[]>`
+        SELECT DISTINCT ON (trajectories.taxi_id)
+            trajectories.taxi_id,
+            taxis.plate,
+            trajectories.date,
+            trajectories.latitude,
+            trajectories.longitude
+        FROM "taxis"
+        JOIN "trajectories" ON taxis.id = trajectories.taxi_id
+        WHERE
+            trajectories.date = (
+                SELECT MAX(trajectories.date)
+                FROM trajectories
+                WHERE trajectories.taxi_id = taxis.id
+            );
+    `
+}
