--- src/routes/latestRoutes.ts
@@ -1,8 +1,0 @@
-import { Router } from "express";
-import { getLatestController } from "../controllers/latestController";
-
-const router = Router();
-
-router.get('/trajectories/latest', getLatestController);
-
-export default router;
