+++ src/routes/trajectoriesRoutes.ts
@@ -1,8 +1,9 @@
 import { Router } from "express";
-import { getTrajectoriesController } from "../controllers/trajectoriesController"
+import { getTrajectoriesController, getLatestController } from "../controllers/trajectoriesController"
 
 const router = Router();
 
 router.get('/trajectories', getTrajectoriesController);
+router.get('/trajectories/latest', getLatestController);
 
 export default router;
